feat(clip): add Dispatcher.ErrorHandling policy
The behavior is still backwards compatible, since the default
zero-value is ContinueOnError. Yet, with this change, it's now
possible for a DispatcherCommand to use the same policies we
use for FlagSet. This change is ergonomic, because now one can
specify ExitOnError and have consistent behavior.

The motivation for this diff is making the behavior more
consistent between FlagSet and DispatcherCommand. I did spend
some time investigating this, ahead of authoring the core
of this diff. So, we also include documentation changes and
"no functional" change simplifications in here.