feat(nparser): new parser implementation
This diff introduces a new incompatible parser implementation.

The core algorithm is the same, save for the fact that the new
implementation allows `--output --` when `--output` requires
an argument, which the old parser does not allow.

The new parser have a much stronger definition of the type
of supported options, which runs along two dimensions:

1. whether the option is groupable or standalone

2. whether the option takes no argument, a mandatory
argument, or an optional argument.

Additionally, in the new parser, each option is bound to
its own prefix, while it is still possible to parse for
multiple prefixes. Therefore, `+verbose` and `--verbose`
are not possible anymore together. One needs to choose
a specific prefix to use.

We also incorporated support for positional arguments
checking into the parser, rather than delegating it
to the upper layers.

We also extended support for custom help flags. Rather
than hardcoding the help flags, we now have a kind of
options that are parsed even before scanning and cause
an early return through a direct match. This allows
custom help flags as well as using the same functionality
for potentially other purposes.

We also refeactored the parsing algorithm to use deques
rather than mainupulating slices explicitly, which
should make it a bit more intuitive.

We also modified the permutation algorithm to use
sorting rather than slices manipulation.

Also, the new parser has a declarative approach and it
will convert what the user has declared in terms of
options and separators into internal maps (in a separate
object, so the parser does not mutate itself).

This design led to a very funny bug. A parser that has
been configured without any option potentially will
parse anything as positional arguments. Yet, this does
not seem desirable, since throwing `--unknown` at it
will lead to an "all good" result. So, in the interest
of making an empty parser slightly more useful, we've
decided to have it default to GNU conventions when
it is empty. (Which honestly seems fair.)