feat(nflag): new flag package
This commit adds the `nflag` package. It is a rewrite of `flag` using
the `nparser` engine (as opposed to `parser`, used by `flag`).

Apart from changing the underlying engine, `nflag` also improves the
overall usability. `FlagSet` uses public fields and documents how they
are initialized. This feels more idiomatic than having a bunch of
getters and setters to deal with.

Help support is also smoother. One needs to call `FlagSet.AutoHelp`
to configure the flags used for help. Previously, we used magic
values like `h` and `help`.

As before, the `FlagSet` is mutable and not goroutine safe, which is
a byproduct of borrowing from the stdlib `flag` package. In any
case, command line parsing is not a business bound to using more
than a goroutine, so this seems justified.

A notable feature (or bug) is that we can configure option prefixes
in the `FlagSet`. However, a flagset with no options defined uses
the parser's default of assuming GNU style options when there's no
option defined. This happens because the parser only learns about
prefixes through options. In any case, this does not feel like a
huge bug (to have a default convention in a product that allows to
specify also different kind of flags).