BREAKING CHANGE(clip): rework reordering and use nflag
This commit switches clip to use nflag and reworks the reordering
algorithm to include the following fixes:

1. require knowing option prefixes for scanning otherwise the
whole process of reordering is too brittle;

2. crucially, stop scanning when we see the separator;

3, stop at the first match otherwise something like this:

```
command subcommand1 --output subcommand2
```

would lead to an ambiguous command line, which we don't want.

Additionally, switch clip to use nflag.

This is a breaking change because reordering is now
conditioned on knowing the option prefixes.