refactor(nflag)!: allow for mechanical refactoring from pflag
The native semantics of this package differs from the native semantics
of flag and pflag with respect to method names and arguments.

The key difference here is that our Foo and FooVar methods follow the
pflag convention of providing long and short flag names, however in
the pflag package, the names would have a `P` suffix.

Additionally, unlike pflag, we do not supply a default value to
the actual function arguments. The original reason for doing this
in nflag was the observation that doing this for `Var` methods
seems silly, since you probably want a pattern where you initialize
a structure with meaningful values and then you don't want your
parsing methods to re-initialize these values again.

However, considering that we would like to provide people with a
reasonable upgrade path from pflag, we probably want to avoid
methods with the same name and a conflicting semantics.

To avoid this issue, proceed as follows: append `Flag` to the
previous method names without changing semantics. This ensures
that we can mechanically upgrade for this diff without any
functional change, but, crucially, now there are no conflicts
with the existing spf13/pflag names and semantics anymore.

BREAKING CHANGE: nflag.FlagSet methods that create flags now
have a `Flag` suffix to distinguish them semantically from
otherwise namesake methods in spf13/pflag.