fix(nflag): bind modified to value, not to flag
We were previously setting modified on a per-flag basis. However, with,
say, `-c, --count <value>` to do not care of whether `-c` OR `--count` has
been modified. We care, instead, about whether the value was modified.

So, refactor the code so that modified is a property of the
value and make sure we add a single value instance when creating
multiple flags (e.g., `-c, --count`).

While there, add tests to ensure we don't break modified in the future.